// src/Helpers/Enums.tsx
export enum StorageKey {
    user = "user",
    language = "language",
    refreshToken = "refreshToken",
    accessToken = "accessToken",
    signupForm = "signupForm",
    email = "email",
    roles = "roles",
}

export enum Category {
    Electronics = "electronics",
    Jewelery = "jewelery",
    MenSClothing = "men's clothing",
    WomenSClothing = "women's clothing",
  }

export enum Sorts {
    byPrice = "SORT_BY_PRICE",
    byRate = "SORT_BY_RATE",
    byCount = "SORT_BY_COUNT"
}